fix: consistency in api file. Rename base folder into shared components
