+++ src/pages/auth/register.tsx
@@ -6,8 +6,8 @@ import { useRouter } from "next/router";
 import { z } from "zod";
 
 import AuthHeading from "@/components/AuthHeading";
-import BaseButton from "@/components/base/BaseButton";
-import Input from "@/components/base/Input";
+import BaseButton from "@/components/shared/BaseButton";
+import Input from "@/components/shared/Input";
 import showToast from "@/utils/showToast";
 import { signUp } from "../api";
 
@@ -97,6 +97,7 @@ export default function Register() {
                 error={errors.password?.message}
                 type="password"
                 id={"password"}
+                showPassword={true}
                 disabled={isLoading}
               />
               <Input
@@ -106,6 +107,7 @@ export default function Register() {
                 type="password"
                 id={"passwordConfirmation"}
                 label="Repeat Password"
+                showPassword={true}
                 disabled={isLoading}
               />
             </div>
